Fix(printk): %llu printed fiction, and the block log believed it
vsnprintf never supported 'll'.  It consumed the second 'l' and fell
into the perfectly valid 'u' case, so the guard written to catch
unsupported conversions never fired.  A 64-bit argument occupies two
varargs slots and is 8-byte aligned, so reading one 32-bit slot did not
truncate the value — it desynchronised every conversion after it.

The cost was a log line that looked like a measurement:

  [BLOCK] sda1: start=4173683 size=1 sectors type=0x00

None of those three numbers were the arguments.  The line was read as
evidence that the MBR parser was broken, which sent the search into a
parser that was correct all along.  Its own code proves it: the loop
skips entries whose type is 0, so a line printed at all cannot describe
type 0x00.

Make the parser say 'll' is unsupported by printing the specifier
itself, and drop the 64-bit casts at the twelve call sites.  Every one
of them prints a sector count, and 32 bits covers 2 TB.  There is no
libgcc here, so a decimal 64-bit conversion would need a hand-written
divide; nothing needs one yet.

Now reads, on the same card that produced the line above:

  [BLOCK] sda1: start=8192 size=61059072 sectors type=0x0c

+++ nothan-kernel/drivers/block/genhd.c
@@ -146,10 +146,10 @@ static int parse_mbr(struct gendisk *disk)
 		disk->part[p].valid      = 1;
 		disk->nr_parts++;
 
-		printk("[BLOCK] %s%d: start=%llu size=%llu sectors type=0x%02x\n",
+		printk("[BLOCK] %s%d: start=%u size=%u sectors type=0x%02x\n",
 		       disk->disk_name, i + 1,
-		       (unsigned long long)e->lba_start,
-		       (unsigned long long)e->lba_size,
+		       (unsigned int)e->lba_start,
+		       (unsigned int)e->lba_size,
 		       (unsigned int)e->type);
 	}
 
@@ -222,8 +222,11 @@ int add_disk(struct gendisk *disk)
 	}
 	disk_table[slot] = disk;
 
-	printk("[BLOCK] %s: registered, capacity=%llu sectors\n",
-	       disk->disk_name, (unsigned long long)disk->capacity);
+	/* capacity is u64, but printk has no 64-bit conversion.  A 32-bit
+	 * sector count covers 2 TB, which is past what this MMC controller
+	 * addresses, so the cast loses nothing that could reach here. */
+	printk("[BLOCK] %s: registered, capacity=%u sectors\n",
+	       disk->disk_name, (unsigned int)disk->capacity);
 
 	/* Whole-disk cdev */
 	register_disk_cdev(disk, slot, 0);
